+++ Answers/lecture01/Kadai1.java
@@ -1,6 +1,6 @@
 package lecture01;
 
-public class Kadai01 {
+class Kadai01 {
     public static void main(String[] args) {
         System.out.println("xxxxxxxxx");
     }
