+++ Answers/lecture01/Kadai2.java
@@ -1,6 +1,6 @@
 package lecture01;
 
-public class Kadai02 {
+class Kadai02 {
     public static void main(String[] args) {
         int studentNumber = 1111111;
         System.out.println("B" + studentNumber + " xxxxxxxxx ");
