+++ Answers/lecture02/Exercise2_2.java
@@ -1,12 +1,11 @@
 package lecture02;
 
-public class Exercise2_2 {
+class Exercise2_2 {
     public static void main(String[] args) {
         Car car = new Car();
-        GasStation gasStation = new GasStation();
 
         car.run();
-        gasStation.refuel(car);
+        GasStation.refuel(car);
         car.run();
     }
 }
