+++ Answers/lecture02/GasStation.java
@@ -1,8 +1,7 @@
 package lecture02;
 
 public class GasStation {
-
-    void refuel(Car car) {
+    public static void refuel(Car car) {
         int fuel = 20;
         car.fuel += fuel;
         System.out.println("給油したことにより、fuelが " + fuel + " 増えました。");
