+++ Answers/lecture02/GasStation03.java
@@ -2,7 +2,7 @@ package lecture02;
 
 public class GasStation03 {
 
-    void refuel(Car03 car) {
+    public static void refuel(Car03 car) {
         int fuel = 20;
         car.fuel += fuel;
         System.out.println("給油したことにより、fuelが " + fuel + " 増えました。");
