+++ Answers/lecture02/Tire.java
@@ -3,7 +3,7 @@ package lecture02;
 public class Tire {
     int size;
 
-    Tire(int size) {
+    public Tire(int size) {
         this.size = size;
     }
 }
